pmix_server_get: avoid creating a local tracker for existing keys
This commit avoids creating a local tracker for existing keys.
A local client may send a get request concurrently with a commit
request from another client, but the server may have processed the
commit request earlyer than the get request. In this case, we create
a local tracker for possibly existing keys that are added with the
completed commit request. Thus, the get request will be pended in
tracker and will be deffered. This scenario is possible when the
non-fence commit-get scheme is used and when the peer GDS
component is `dstore`.

Signed-off-by: Boris Karasev <karasev.b@gmail.com>