Enable search for available servers
Tools/debuggers need a mechanism by which they can discover the servers
available to the caller on the local node. Provide a query attribute
plus a couple of corresponding attributes for returning the information.
Extend the information placed in the rendezvous to include useful things
like the pid of the server, the uid/gid it is running under, and the
date/time it started.

Signed-off-by: Ralph Castain <rhc@pmix.org>