Point to the correct file that wasn't found
When doing a dlopen, we apparently strip the extension from the actual
filename to identify the component it relates to, and then add accepted
extensions back to the filename when checking for file existence prior
to loading. Unfortunately, some 3rd party packages (e.g., debuggers)
take our libraries, do a little magic, and then output a copy of the
results into our lib directory - only with a new suffix. So we might get
something like "ourcomponent.so.dbg". We then strip the ".dbg" and add
".so" to it again - and complain that "ourcomponent.so is not found".

Fix the error message so it reports the name of the file it actually
couldn't find, not the name of the component, so it doesn't totally
confuse people.

Signed-off-by: Ralph Castain <rhc@pmix.org>