Begin implementation of Slingshot plugin
Fix the check for libjansson so it properly handles the
case where the library is located under the typical place
under the --with-jansson directory.

Begin work on creating the Slingshot cost "matrix".
Create a JSON generator for test Slingshot topology files.
Create a pnet/sshot component that reads the topology
file and generates an update to the scheduler with the group
memberships. Provide a test server/daemon pair that emulates the
eventual behavior.

Signed-off-by: Ralph Castain <rhc@pmix.org>