build: Support hwloc and libevent cobuild
Add ability for PMIx to find hwloc and libevent packages which are
built at the same time as PMIx, such as third party dependencies
in an MPI package.  This is a slight change on the existing
embedded PMIx support, in that it does not assume that PMIx itself
is being built embedded in another library, only that some of its
dependencies may be built at the same time as PMIx.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>