Delete dead code
The static availsys array is initialized with doubles where the underlying
struct type uses uint64_t resulting in compiler warnings.
Removing all related code using this availsys array as it is only used in
a single path that has no side effect to any other part in the code base.

Signed-off-by: Christoph Niethammer <niethammer@hlrs.de>