Fix cross-version for v3.1.6+
We hadn't really planned on releasing anything beyond rel=5 in the 3.1.x
series, so we hard-coded a check for 3.1.5 in client get. This must be
updated to allow 3.1.6 (plus any future releases in the v3.1.x series)
to pass cross-version tests when the v3.1.x release is acting as the
server.

Signed-off-by: Ralph Castain <rhc@pmix.org>
(cherry picked from commit 56ccc1ec37c906af7659dd2021657d6c7f053eb6)