Complete slingshot scaffold
Cleanup checks for curl and jansson support. Ensure they properly get
included in the common/sse component and cleanup a few errors there.
Extend the pcompress framework to compress blocks of data instead of
just strings. Allow the pnet/sshot component to build when curl and
jansson support is available and the user specifically requests it to be
built by adding --with-slingshot. Extend the regex support to include
the "raw" option. Add PMIX_COMPRESSED_BYTE_OBJECT data type.

Signed-off-by: Ralph Castain <rhc@pmix.org>