Protect the socket/bind call and avoid unnecessary key waits
When performing a get operation that goes up to the server, check to see
if the key exists in some other scope and return the correct error if it
does. Check for that error and don't "wait" for the data to appear as it
will never do so if it is already posted to a scope outside the range of
the caller.

Protect the socket/bind call when running on Mac.

Update test_job_fence so only one proc tries one time to find a
nonexistent key - helps reduce the runtime of the test

Signed-off-by: Ralph Castain <rhc@pmix.org>