Cleanup indirect debugger launch against mpirun
There are some differences between a debugger working against a tool that directs an RM to launch the job (e.g., prun) versus working against the launcher itself (e.g., mpirun). Allow servers to drive a connection attempt so the server can connect back to the tool. Have the server respect the wait-for-debugger-release request. Aggregate the app's waiting-for-debugger events so the host isn't getting hit with lots of individual ones.

Signed-off-by: Ralph Castain <rhc@pmix.org>