Silence unnecessary warning
If a client sees a "lost connection" error, there is no point in trying to send it to the server as the connection to the server is what has been lost! Also protect against the case where the connection is lost in mid-communication - we flag that by returning a zero-byte message, so don't try to unpack it.

Signed-off-by: Ralph Castain <rhc@pmix.org>