Fix fence timeout
The PMIx library can only test timeout for local completion - we
cannot continue the timer once we pass the operation to the host
as we would have conflicting timers running. Correctly have the
tracker (and not the individual proc entry's) event carry the
timeout. Fix reuse of a status variable to ensure that all
participants receive the "timeout" status code.

Signed-off-by: Ralph Castain <rhc@pmix.org>