build: Remove options around mca direct calling
The ability to directly call an mca component as if it was a
library function was originally added for small, high
performance code segments, like MPI interface calls
implemented over an MPI-centric network interface (Portals,
UCX, MX, etc.).  It requires significant work in the framework,
all components in the framework, and all callers of the
framewrok, and none of the frameworks in PMIx have that
infrastructure (and likely never will).  Therefore, we are
removing all the direct call references in the code.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>