Reject ambiguous connection options
If a tool finds multiple servers that it could connect to, then it
shall return an error indicating that the user must provide more
information so a specific server can be selected. Otherwise, it can
be a somewhat random selection. Multiple connection files are permitted,
but we check to see if all point to the same server. If not, then
report the ambibuity and ask for clarification.

Signed-off-by: Ralph Castain <rhc@pmix.org>
(cherry picked from commit 3a659c0cd9b64c0379226f643ed71a855d96a207)