Add a wildcard nspace definition
We were using an empty pmix_nspace_t as "invalid" and "wildcard".
However, this can lead to confusion between nspace objects that
haven't yet been defined and those situations when we really
do want to check against all nspaces. So define a "wildcard" to
be a pmix_nspace_t that has nothing but '*' characters in it
(retaining the NULL to end the string).

Also update the clang format file and script

Signed-off-by: Ralph Castain <rhc@pmix.org>