Some mods/cleanup of debugger definitions and handling
The host/debugger needs a way to communicate which ranks are to
be debugged - eg., which are to be stopped on exec. We had been
assuming that it would be all of them as that is the typical
case, but that isn't a restriction. Modify the debugger attrs
to pass a rank (or an array of ranks) instead of just "bool".
Allow the app to indicate where they are stopped, should they
choose to do so. Replace `WAIT_FOR_NOTIFY` with `READY_FOR_DEBUG`
as release by notification is probably the least-used use-case.

Signed-off-by: Ralph Castain <rhc@pmix.org>
(cherry picked from commit e842c027a093219295b297dfa73681cde3a830b2)