Transfer stdout/err formatting to PMIx
Transfer the PRRTE output formatting code to the PMIx library
so that tools can also provide a similar capability. This
includes the ability to output XML format, timestamps,
and tagged output as well as the ability to specify a
file format and/or directory where the output is to be
written.

Please see the OpenPMIx wiki page for explanation of
this feature and how to use it.

Signed-off-by: Ralph Castain <rhc@pmix.org>
(cherry picked from commit ebf361c796b40770d1091ef394d50a859bcd3ac3)