common/sse: Ensure library links against libcurl
Unify on a single naming convention for the library produced for
the common/sse component, with libmca_common_sse.la.  Unlike
traditional components that are actual DSOs, common components
aren't dynamically loaded; they're dynamically linked by other
loaded components and are stand-alone shared libraries, so the
naming convention that is necessary for other components is
not necessary for common components.

This solves a bug with the common/sse component not being able
to find libcurl symbols when built as a static component.  The
*_LIBADD rule was only for the dynamic build mode and was missing
for the static build mode.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>