common/dstore: Fix inconsistent Makefile.am
The Makefile.am for common/dstore was inconsistent in the
flags set for the static and dynamic build cases.  While
this was not causing any problems today, it could lead
to the problems seen with common/sse on the master branch.

This change is similar to 936bf79 on master, in that it
unifies on a single naming convention for the library produced
by common components.

At the same time, remove some unneeded flags, like the LTDLINC
CPPFLAGS, which are not used in this component.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>