Do not call client_finalized for tools, report lost tool connections
Both clients and tools send the PMIX_FINALIZE_CMD to the server when
finalizing, but we don't want the server calling the host's
client_finalized function for tools. A host can only learn of a tool's
departure by receiving a "lost connection" event, so be sure to always
generate one when a tool departs.

Signed-off-by: Ralph Castain <rhc@pmix.org>