Some cleanup of IOF output
No longer directly output stdout/stderr - add them to the event
list for output. This helps us deal with the case where the std
file descriptors are getting blocked/metered. Add a counter to
track the number of times we attempt to output to a given sink
and abort if we exceed that count - helps avoid deadlock where
we simply cannot get the output to go. Also, don't try to "join"
the progress thread when we stop it as this can hang if we are
in an event we cannot get out of.

Signed-off-by: Ralph Castain <rhc@pmix.org>