Properly read/output stdout/err from a fork/exec'd child
If a tool calls "spawn" while not connected, we will fork/exec
the specified app as a child of the PMIx library. In this case,
the library takes responsibility for reading, forwarding, and/or
outputting any stdout/err from that child.

Signed-off-by: Ralph Castain <rhc@pmix.org>
(cherry picked from commit d7587471ab2c63051e516646472e1d43e8cdafc3)