pmix.m4: abort if gcc is <= v4.8.1
This is a port from the corresponding
https://github.com/open-mpi/ompi/pull/9398.

OpenPMIx does not compile with GCC 4.4.7 (the default GCC that ships
in RHEL 6).  Open PMIx *does* compile with GCC 4.8.5 (the default GCC
that ships with RHEL 7).  The Open MPI community regularly tests Open
MPI (including PMIx) with GCC 4.8.1, so we know that works, too.

Hence, abort configure if GCC is used and GCC is < v4.8.1.

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>