Cleanup the test_v2 area
Avoid multiple definitions of a couple of variables declared
in test_common.h by making them "extern", instancing them
in test_common.c. Silence warnings on Mac due to use of
"-no-install" option in Makefile - it is "no-fast-install"
on the Mac. Although cute, we don't support Windows so let's
avoid the warnings from the "sleep" function.

Update ignores as well.

Signed-off-by: Ralph Castain <rhc@pmix.org>