start on squash of unused vars in examples dir
NOTE: These are supposed to be more stand-alone so
copied the PMIX_HIDE_UNUSED_PARAMS() macro over
to a new examples.c (common) file as EXAMPLES_HIDE_UNUSED_PARAMS().

The problem is that some codes duplicate bits in examples.h
and cause redefine errors if you use that header.  So in this
verison we just do a `(void)var` with comment for searching.

Once we can cleanly include examples.h in all examples w/o redfines,
we can just use the macro.

Signed-off-by: Thomas Naughton <naughtont@ornl.gov>