Cleanup configuration logic
We still weren't dealing with the scenario where there is
a copy of HWLOC in the standard location, but we pass our
own path to a different HWLOC install that we want to use.
Problem turned out to be the common mistake we keep rolling
back to making - AC_CHECK_HEADERS _always_ looks in the default
paths in addition to whetever you may add. Ditto for its
AC_SEARCH_LIBS companion.

So you have to be more careful about handling this situation. We
periodically seem to forget and rollback the configure logic
into this error, so try to provide a big comment explaining
why you can't do it.

If I felt like taking the time to search for it, I could swear
I could just have copied that comment from prior times...

<shrug>

Also, cleanup a few of the summary lines at the end of configure,
and a utility to convert HWLOC XML between versions 2 and 1.

Signed-off-by: Ralph Castain <rhc@pmix.org>