Fix the network support components
When we allocate information to support a particular fabric,
we pack that info into a buffer and then transport the "blob"
to the backend for unpacking and processing. During the unpack
operation, we tried to be clever and simply assign the blob
into the various buffer fields.

Bad idea - missed a critical one and that caused the unpack
to fail. Instead, define a suitable macro for this purpose
to ensure that ALL the required buffer fields get set.

Also, OPA fabric is device type 0x208 - correct the typo.

Thanks to Howard Pritchard and Michael Heinz for identifying
the problem.

Signed-off-by: Ralph Castain <rhc@pmix.org>