Remove curl/jansson default search assumption
Remove the assumption that the default search path for headers
and libraries is /usr/include and /usr/lib,/usr/lib64.  Ubuntu
20.04 on ARM uses /usr/lib/aarch64-linux-gnu for most of its
libraries (which is in the compiler default search paths).  MacOS
has a default search path for includes that is inside the XCode
package.

I can't find any platforms on which curl.h is installed in
include instead of curl/curl.h and the pathing setup for
trying to handle both cases was broken (it would result in
a broken library search unless the user also gave the libdir),
so simplify that logic as well.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>