Fix a number of warnings and cleanup a few things
Clarify that the inputs to pmix_ptl_base_split_and_resolve are
not to be disturbed - one is an MCA param string and the other
is a constant string. Fix an uninitialized variable (match_count).
Fix the show_help messages - missing an input parameter to the
message. Use the --enable-devel-check and fix some remaining
unused variables.

Signed-off-by: Ralph Castain <rhc@pmix.org>