Clean up libevent/libev selection logic
The primary change in this patch is to properly enforce the
interlock between libev and libevent selection logic.  Only one
of the two packages can be used by a PMIX build, but that was
not previously enforced (only the check that the user did not
ask for both was enforced).  There are also some cleanupt s around
prints when a check function knows that package will not be
selected.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>