Remove pkg-config dependency list
Replace the dependency list in pkg-config with the list of required
libraries.  This is not a long term solution, but we need similar
infrastructure to the wrapper LIBS for PC files, and that is a bigger
change.  We also need to do something about libevent; declaring a
requires on libevent brings in -levent, which causes problems since
pmix is linked against event_core (sigh; libevent made a mess here).

Signed-off-by: Brian Barrett <bbarrett@amazon.com>