Remove unneeded atomics code
PMIx is a single-threaded code base. We are either
in the user's thread, or we are in the PMIx progress
thread. Objects are not allowed to be accessed by
both threads at the same time. Therefore, we don't
need all that atomics code - we only need the rmb/wmb
functions to handle the threadshifts.

Use pthread_mutex protection around the object
reference count updates.

Signed-off-by: Ralph Castain <rhc@pmix.org>
(cherry picked from commit a98fbbc8d033169dd9c5d265905591e0853c7068)