Be defensive against string literals in env
Add two new functions to the pmix_environ code: pmix_getenv() and
pmix_environ_merge_inplace().  The first is an implementation of
getenv() with a second parameter to be explicit about the environment
to search.  The second adds the contents of an environment-style
key=value argv array into another environment, not overwriting
the values of existing keys.  The search does not modify
the added environment keys, protecting PMIx from string literals
or other read-only data in environ when building application
environments.

Update the pfexec code to use pmix_environ_merge_inplace().

Add a test to make sure the new code works (and test some of the
existing code, although there is still room for improvement).

Signed-off-by: Brian Barrett <bbarrett@amazon.com>