Better handle flags in wrapper & pkg-config
Fix a bug in the wrapper compiler where we were adding a bare directory
to the wrapper CPPFLAGS, which was committed as part of 790fd08.

Reorder the wrapper-data file so that the -I{includedir} and -L{libdir}
are first in the CPPFLAGS and LDFLAGS (respectively) and remove the
duplicate -I{includedir} that was caused by it being added to the
WRAPPER FLAGS.

Add all the wrapper preprocessor flags to the pmix.pc file.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>