Silence warnings in pmixt_validate_predefined().
Silence "reading 512 bytes from a region of size X" warnings generated
by GCC 11.2.0. This may be a spurious warning, but once that is
generated nonetheless. The problem seems to come from a function
definition with a static length char buffer (pmix_key_t) with call-sites
passing constant strings that are strictly smaller than the size of
pmix_key_t. This usage seems atypical in the codebase, so perhaps there
is a better way to silence these particular warnings.

A similar warning is generated at a call-site in
test/test_v2/server_callbacks.c at line 323 involving a call to
PMIx_server_register_nspace(), but this one is stickier, so someone else
should look into that one. Note: changing the string constant to a char*
variable does not silence this one. Pesky, pesky.

Signed-off-by: Samuel K. Gutierrez <samuel@lanl.gov>