Provide a "hook" for PRRTE to init a small amount of PMIx
Now that PRRTE directly depends on PMIx classes and frameworks,
we need to ensure that some small portion of PMIx is setup
prior to PRRTE attempting to use it. Create a new "pmix_init_util"
for this purpose. Need to isolate that prototype in its own
include header to avoid pulling in a bunch of other stuff.

The PMIx interface framework was the immediate problem as PRRTE
tried to traverse a list prior to it being constructed.

Signed-off-by: Ralph Castain <rhc@pmix.org>