Properly handle queries of tools
If a tool receives a query, it may (and is likely) not be able
to find the data being requested. For example, if a tool requests
the proc table for a job that another tool spawned, then that
request has to be relayed to the RM for an answer - and then the
answer has to be relayed back to the original requestor.

Fix a problem in simptest where a threadshift left a value on
the stack, which could subsequently be lost.

Signed-off-by: Ralph Castain <rhc@pmix.org>