Prohibit Python bindings with non-shared lib builds
Cython takes issue with purely static builds, so error
out if the user requests the Python bindings but has
specified --disable-shared.

Also fix a problem with the "canonicalize_path" function
in configure - new versions of MacOS (i.e., starting with
Mojave) do _not_ have a /usr/bin/python, but do typically
have a "python" in the path. However, this python is often
v3, so make the syntax "futureproof" to handle the situation

Signed-off-by: Ralph Castain <rhc@pmix.org>