build: Use MCA env management for common/sse
Update the Curl and Jansson checks to not AC_SUBST their compile/link
flags, but instead have the common/sse component build the flags
variables using the output of the Curl/Jansson tests, but named
using the standard component convention.  This allows the MCA system
to add the flags to the wrapper flags without a bunch of extra mucking
about.

This fixes an issue where -lcurl / -ljansson weren't propogated to
the wrapper compilers if common/sse is built as part of libpmix.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>