Remove incorrect object release
Cannot release the "lg" object as it might be used again
in the loop. Object will be released when the caddy is
released.

Cleanup the simpdmodex test a bit so that all outputs are
properly clumped when run thru "sort" - makes it easier
to debug.

Signed-off-by: Ralph Castain <rhc@pmix.org>
(cherry picked from commit ea4ad6114daf3e4fdbae9d328022117c8af4474c)