Properly deal with delayed local get requests
Current code only supported delayed "get" requests
(i.e., a request for data not yet received) from
remote procs. If a local proc requested data from
another local proc, and the posting hadn't already
been done, then we would wait - but the retrieval
after waiting was locked to PMIX_REMOTE scope, and
so the data would never be found.

Also fix the return code so that we get the original
issue ("timeout") and not a bogus "not found".

Increase the flexibility of the simpdmodex test to
allow passing timeout, sleep, and async flags on
the cmd line.

Adjust the tests to not error on timeout when expected
to fail

Signed-off-by: Ralph Castain <rhc@pmix.org>
(cherry picked from commit 30ee3862aeea3d0dde714d9b2dfdd9163585c3b7)