Fix dmodex operation on local host
Pass the scope to pending_resolve so we check the local
scope when appropriate instead of having the function locked
to the remote scope case. This was a change that somehow didn't
get backported from master.

Also ensure that we check the peer's GDS component and not
just our own when looking for requested data. On master, there
currently is only one storage system, so it doesn't matter.
However, that isn't the case on prior release branches, and
data can be stored in the peer's component.

Update the simpdmodex test to track master

Signed-off-by: Ralph Castain <rhc@pmix.org>