Ensure we check peer's gds for data
When resolving data requests that may have been delayed waiting
for the source to commit the requested data, be sure to check
both the server's _and_ the peer's GDS component for the data.
While master currently only has one component, that may not
be true in the future, and it is possible that the committed
data was stored on the peer's component (and not the server's).

Signed-off-by: Ralph Castain <rhc@pmix.org>