Correctly handle qualified values
Allow the user to put/get values that are qualified - i.e., a process
might store multiple values for the same key, each value being
qualified to specify when that value (as opposed to the others) shall
be returned. When a user issues a "get" for the key, return the
value which matches the provided qualifiers - a "get" without
qualifiers will only return a value that was "put" without qualifiers.

Signed-off-by: Ralph Castain <rhc@pmix.org>