Streamline operations
Stop storing some data multiple times. Cleanup a bug in
hash store_job_info that stored the proc data as a block
instead of pulling out the individual data elements - this
resulted in simply overwriting the block and losing the
proc-level data.

Provide a new attribute for storing a tool or server's own
URI separate from the URI of the server to which they are
connected.

Signed-off-by: Ralph Castain <rhc@pmix.org>