Support colocation of processes
When spawning a new job, support the case where the user
wants to spawn the new procs at locations where current
procs are executing. Provide for both specification of
the number of new procs to start on each node where the
current procs are executing, or the number of new procs
to start for each current proc that is executing.

Note that these are not required to be daemons - they can
be any application code, and can have any desired ranking
and/or binding policy. Mapping policy, however, will be
ignored as it is overridden by the "colocate" directive.

Signed-off-by: Ralph Castain <rhc@pmix.org>
(cherry picked from commit 137cc3d6b3c99065bb55633f6cf0568aecdc34d4)