Cleanup the show_help changes a bit
It is true that pmix_status_t is an int, but we always
explicitly return pmix_status_t anyway (just in case it
someday changes, which would be hard to believe). Add a
"norender" API that PRRTE uses in only one place, but easier
than trying to figure out a way around it.

Signed-off-by: Ralph Castain <rhc@pmix.org>