Modify the key-index support
Utilize a pointer array instead of a hash table for
faster retrieval. Let the key's index correspond to
the index in the array. Append user-defined keys to
the standard ones. Use that relative location to reduce
the search space/time when retrieving an index from
a key.

Add some missing static initializers for internal
structures.

Signed-off-by: Ralph Castain <rhc@pmix.org>