Require flex only when keyval_lex.c is not provided
We currently require flex whenever we are in a Git clone, but that
really isn't the requirement. We need flex whenever the flex output
files are not present - otherwise, you can build just fine. So open
things up a bit by tying the flex requirement to the actual one
(i.e., that the flex output file exist).

Signed-off-by: Ralph Castain <rhc@pmix.org>