Some initial valgrind cleanup
Provide a way to release the
pmix_value_t and pmix_data_array_t storage. Ensure
we release the pmix_job_t tracker in the gds/hash
component when the nspace is deleted.

Resolve memory corruption problem associated with passing
the HWLOC topology into the PMIx server from the host. This
was being copied during the "info_list" processing, and then
being deleted when the resulting data array was released - leaving
the server to continue accessing the free'd topology tree.

Signed-off-by: Ralph Castain <rhc@pmix.org>